+++ app/models/views/reports/hmrc_ocmc_data_export.rb
@@ -3,6 +3,7 @@ module Views
   module Reports
     class HmrcOcmcDataExport
       require 'csv'
+      include OcmcExportHelper
 
       def initialize(start_date, end_date, office_id)
         @date_from = format_dates(start_date)
@@ -25,14 +26,6 @@ module Views
       end
 
       private
-
-      def data
-        @data ||= build_data
-      end
-
-      def build_data
-        ActiveRecord::Base.connection.execute(sql_query)
-      end
 
       # rubocop:disable Metrics/MethodLength
       def sql_query
@@ -194,24 +187,6 @@ module Views
         from = date_formatted(date_range)[:from]
         to = date_formatted(date_range)[:to]
         "#{from} - #{to}"
-      end
-
-      def children_age_band(value, attr_key)
-        return nil if age_bands_blank?(value)
-        hash_value = YAML.safe_load(value, permitted_classes: [Symbol])
-        if attr_key == :children_age_band_one
-          (hash_value[:one] || hash_value['one'])
-        elsif attr_key == :children_age_band_two
-          (hash_value[:two] || hash_value['two'])
-        end
-      end
-
-      def age_bands_blank?(value)
-        return true if value.blank?
-        hash_value = YAML.safe_load(value, permitted_classes: [Symbol])
-        hash_value.keys.select do |key|
-          key.to_s == 'one' || key.to_s == 'two'
-        end.blank?
       end
 
     end
