+++ app/models/views/reports/ocmc_data_export.rb
@@ -2,6 +2,7 @@ module Views
   module Reports
     class OcmcDataExport
       require 'csv'
+      include OcmcExportHelper
 
       def initialize(start_date, end_date, office_id)
         @date_from = format_dates(start_date)
@@ -30,14 +31,6 @@ module Views
         csv_row['Age band under 14'] = children_age_band(row['Age band under 14'], :children_age_band_one)
         csv_row['Age band 14+'] = children_age_band(row['Age band 14+'], :children_age_band_two)
         csv_row
-      end
-
-      def data
-        @data ||= build_data
-      end
-
-      def build_data
-        ActiveRecord::Base.connection.execute(sql_query)
       end
 
       # rubocop:disable Metrics/MethodLength
@@ -92,25 +85,6 @@ module Views
         AND applications.state != 0 ORDER BY applications.created_at DESC;"
       end
       # rubocop:enable Metrics/MethodLength
-
-      def children_age_band(value, attr_key)
-        return nil if age_bands_blank?(value)
-        hash_value = YAML.safe_load(value, permitted_classes: [Symbol])
-        if attr_key == :children_age_band_one
-          (hash_value[:one] || hash_value['one'])
-        elsif attr_key == :children_age_band_two
-          (hash_value[:two] || hash_value['two'])
-        end
-      end
-
-      def age_bands_blank?(value)
-        return true if value.blank?
-
-        hash_value = YAML.safe_load(value, permitted_classes: [Symbol])
-        hash_value.keys.select do |key|
-          key.to_s == 'one' || key.to_s == 'two'
-        end.blank?
-      end
 
     end
   end
